+++ packages/x-charts/src/constants/index.ts
@@ -11,7 +11,7 @@ export const DEFAULT_MARGINS = {
 export const DEFAULT_AXIS_SIZE_WIDTH = 45;
 export const DEFAULT_AXIS_SIZE_HEIGHT = 25;
 
-export const DEFAULT_ZOOM_BRUSH_SIZE = 30;
+export const DEFAULT_ZOOM_OVERVIEW_SIZE = 30;
 
 // How many pixels to add to the default axis size if that axis has a label
 export const AXIS_LABEL_DEFAULT_HEIGHT = 20;
