+++ scripts/x-charts-pro.exports.json
@@ -149,7 +149,7 @@
   { "name": "DEFAULT_ROTATION_AXIS_KEY", "kind": "Variable" },
   { "name": "DEFAULT_X_AXIS_KEY", "kind": "Variable" },
   { "name": "DEFAULT_Y_AXIS_KEY", "kind": "Variable" },
-  { "name": "DEFAULT_ZOOM_BRUSH_SIZE", "kind": "Variable" },
+  { "name": "DEFAULT_ZOOM_OVERVIEW_SIZE", "kind": "Variable" },
   { "name": "DefaultizedBarSeriesType", "kind": "Interface" },
   { "name": "DefaultizedCartesianSeriesType", "kind": "TypeAlias" },
   { "name": "DefaultizedFunnelSeriesType", "kind": "Interface" },
